backend/requirements.txt
# Core
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.3
pydantic-settings>=2.1.0
email-validator>=2.0.0

# Database
sqlalchemy[asyncio]>=2.0.25
asyncpg>=0.29.0
alembic>=1.13.1

# Auth
python-jose[cryptography]>=3.3.0
authlib>=1.3.0
httpx>=0.27.0
python-multipart>=0.0.6
itsdangerous>=2.0.0

# YouTube
yt-dlp>=2024.11.4

# AI
anthropic>=0.18.1

# Telegram
python-telegram-bot>=21.0

# Background Jobs
apscheduler>=3.10.4

# Cache
redis>=5.0.0

# Security
slowapi>=0.1.9

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.3
